# api/gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Domilike Service
    version: ""
host: localhost:18080
basePath: /api
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
    /cards/{id}:
        get:
            tags:
                - user
            summary: get user
            operationId: user#get
            parameters:
                - name: id
                  in: path
                  description: Card ID
                  required: true
                  type: integer
                - name: Authorization
                  in: header
                  description: JWT token used to perform authorization
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: string
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
...
securityDefinitions:
    jwt_header_Authorization:
        type: apiKey
        description: JWT based authentication
        name: Authorization
        in: header